feat: voronoi図形の面積計算を効率化
